Continuous Action Multi Env Training for Goal-reach using Openai Baselin
Added a Continuous Multi-Env where the Agent gets trained in 3 env simultaneously. At the begging of each episode and Env is randomly selected (Multi-lane, Merge, or Roundabout), populated and a goal is randomly located on the road. The agent must reach the goal taking into account, the direction of the traffic and the road structure while avoiding collisions.
The parking_env was originally developed by Edouard Leurent.